+++ Snake/MainWindow.xaml.cs
@@ -7,7 +7,6 @@ using System.Windows.Controls;
 using System.Windows.Data;
 using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
@@ -28,11 +27,14 @@ namespace Snake
         List<PositionedEntity> snake;
         // яблоко
         Apple apple;
+
+        PoisonedApple papple;
         //количество очков
         int score;
         //таймер по которому 
         DispatcherTimer moveTimer;
-        
+        DispatcherTimer poisonTimer;
+
         //конструктор формы, выполняется при запуске программы
         public MainWindow()
         {
@@ -46,7 +48,11 @@ namespace Snake
             moveTimer = new DispatcherTimer();
             moveTimer.Interval = new TimeSpan(0, 0, 0, 0, 300);
             moveTimer.Tick += new EventHandler(moveTimer_Tick);
-            
+
+            poisonTimer = new DispatcherTimer();
+            poisonTimer.Interval = new TimeSpan(0, 0, 0, 10, 0);
+            poisonTimer.Tick += new EventHandler(poisonTimer_Tick);
+            poisonTimer.Stop();
         }
 
         //метод перерисовывающий экран
@@ -62,56 +68,106 @@ namespace Snake
             //обновляем положение яблока
             Canvas.SetTop(apple.image, apple.y);
             Canvas.SetLeft(apple.image, apple.x);
-            
+
+            //обновляем положение отравленного яблока
+            if (!poisonTimer.IsEnabled && papple != null)
+            {
+                Canvas.SetTop(papple.image, papple.y);
+                Canvas.SetLeft(papple.image, papple.x);
+            }
+
             //обновляем количество очков
             lblScore.Content = String.Format("{0}000", score);
         }
-
+        void poisonTimer_Tick(object sender, EventArgs e)
+        {
+            canvas1.Children.Remove(head.image);
+            head.ChangeImage("pack://application:,,,/Resources/head.png");
+            head.image.RenderTransformOrigin = new Point(0.5, 0.5);
+            canvas1.Children.Add(head.image);
+            poisonTimer.Stop();
+            poisonTimer.IsEnabled = false;
+        }
         //обработчик тика таймера. Все движение происходит здесь
         void moveTimer_Tick(object sender, EventArgs e)
         {
             //в обратном порядке двигаем все элементы змеи
             foreach (var p in Enumerable.Reverse(snake))
-            {
                 p.move();
-            }
 
+            gameover_Check();
+            //проверяем, что голова змеи врезалась в яблоко
+            if (head.x == apple.x && head.y == apple.y)
+                apple_occuried();
+            //проверяем, что голова змеи врезалась в отравленное яблоко
+            else if (papple != null && head.x == papple.x && head.y == papple.y)
+                poison_apple_occuried();
+            //перерисовываем экран
+            UpdateField();
+        }
+        
+        void apple_occuried()
+        {
+            //увеличиваем счет
+            score++;
+            //двигаем яблоко на новое место
+            apple.move();
+            // добавляем новый сегмент к змее
+            snake_grow();
+            if (papple != null)
+            {
+                canvas1.Children.Remove(papple.image);
+                papple = null;
+            }
+            else if ((new Random()).Next(0, 100) < 5 && !poisonTimer.IsEnabled)
+            {
+                papple = new PoisonedApple(snake, apple);
+                canvas1.Children.Add(papple.image);
+            }
+        }
+        void poison_apple_occuried()
+        {
+            poisonTimer.Start();
+            poisonTimer.IsEnabled = true;
+            score += 25;
+            apple.move();
+            canvas1.Children.Remove(papple.image);
+
+            canvas1.Children.Remove(head.image);
+            head.ChangeImage("pack://application:,,,/Resources/phead.png");
+            head.image.RenderTransformOrigin = new Point(0.5, 0.5);
+            canvas1.Children.Add(head.image);
+
+            snake_grow();
+
+            papple = null;
+        }
+        void gameover_Check()
+        {
             //проверяем, что голова змеи не врезалась в тело
             foreach (var p in snake.Where(x => x != head))
-            {
                 //если координаты головы и какой либо из частей тела совпадают
                 if (p.x == head.x && p.y == head.y)
-                {
-                    //мы проиграли
-                    moveTimer.Stop();
-                    tbGameOver.Visibility = Visibility.Visible;
-                    return;
-                }
-            }
+                    gameover();
 
             //проверяем, что голова змеи не вышла за пределы поля
             if (head.x < 40 || head.x >= 540 || head.y < 40 || head.y >= 540)
-            {
                 //мы проиграли
+                gameover();
+
+            void gameover()
+            {
                 moveTimer.Stop();
+                poisonTimer.Stop();
                 tbGameOver.Visibility = Visibility.Visible;
                 return;
             }
-
-            //проверяем, что голова змеи врезалась в яблоко
-            if (head.x == apple.x && head.y == apple.y)
-            {
-                //увеличиваем счет
-                score++;
-                //двигаем яблоко на новое место
-                apple.move();
-                // добавляем новый сегмент к змее
-                var part = new BodyPart(snake.Last());
-                canvas1.Children.Add(part.image);
-                snake.Add(part);
-            }
-            //перерисовываем экран
-            UpdateField();
+        }
+        void snake_grow()
+        {
+            var part = new BodyPart(snake.Last());
+            canvas1.Children.Add(part.image);
+            snake.Add(part);
         }
 
         // Обработчик нажатия на кнопку клавиатуры
@@ -120,16 +176,16 @@ namespace Snake
             switch (e.Key)
             {
                 case Key.Up:
-                    head.direction = Head.Direction.UP;
+                    head.direction = poisonTimer.IsEnabled ? Head.Direction.DOWN : Head.Direction.UP;
                     break;
                 case Key.Down:
-                    head.direction = Head.Direction.DOWN;
+                    head.direction = poisonTimer.IsEnabled ? Head.Direction.UP : Head.Direction.DOWN;
                     break;
                 case Key.Left:
-                    head.direction = Head.Direction.LEFT;
+                    head.direction = poisonTimer.IsEnabled ? Head.Direction.RIGHT : Head.Direction.LEFT;
                     break;
                 case Key.Right:
-                    head.direction = Head.Direction.RIGHT;
+                    head.direction = poisonTimer.IsEnabled ? Head.Direction.LEFT : Head.Direction.RIGHT;
                     break;
             }
         }
@@ -161,163 +217,6 @@ namespace Snake
             UpdateField();
 
         }
-        
-        public class Entity
-        {
-            protected int m_width;
-            protected int m_height;
-            
-            Image m_image;
-            public Entity(int w, int h, string image)
-            {
-                m_width = w;
-                m_height = h;
-                m_image = new Image();
-                m_image.Source = (new ImageSourceConverter()).ConvertFromString(image) as ImageSource;
-                m_image.Width = w;
-                m_image.Height = h;
 
-            }
-
-            public Image image
-            {
-                get
-                {
-                    return m_image;
-                }
-            }
-        }
-
-        public class PositionedEntity : Entity
-        {
-            protected int m_x;
-            protected int m_y;
-            public PositionedEntity(int x, int y, int w, int h, string image)
-                : base(w, h, image)
-            {
-                m_x = x;
-                m_y = y;
-            }
-
-            public virtual void move() { }
-
-            public int x
-            {
-                get
-                {
-                    return m_x;
-                }
-                set
-                {
-                    m_x = value;
-                }
-            }
-
-            public int y
-            {
-                get
-                {
-                    return m_y;
-                }
-                set
-                {
-                    m_y = value;
-                }
-            }
-        }
-
-        public class Apple : PositionedEntity
-        {
-            List<PositionedEntity> m_snake;
-            public Apple(List<PositionedEntity> s)
-                : base(0, 0, 40, 40, "pack://application:,,,/Resources/fruit.png")
-            {
-                m_snake = s;
-                move();
-            }
-
-            public override void move()
-            {
-                Random rand = new Random();
-                do
-                {
-                    x = rand.Next(13) * 40 + 40;
-                    y = rand.Next(13) * 40 + 40;
-                    bool overlap = false;
-                    foreach (var p in m_snake)
-                    {
-                        if (p.x == x && p.y == y)
-                        {
-                            overlap = true;
-                            break;
-                        }
-                    }
-                    if (!overlap)
-                        break;
-                } while (true);
-
-            }
-        }
-
-        public class Head : PositionedEntity
-        {
-            public enum Direction
-            {
-                RIGHT, DOWN, LEFT, UP, NONE
-            };
-
-            Direction m_direction;
-
-            public Direction direction {
-                set
-                {
-                    m_direction = value;
-                    RotateTransform rotateTransform = new RotateTransform(90 * (int)value);
-                    image.RenderTransform = rotateTransform;
-                }
-            }
-
-            public Head()
-                : base(280, 280, 40, 40, "pack://application:,,,/Resources/head.png")
-            {
-                image.RenderTransformOrigin = new Point(0.5, 0.5);
-                m_direction = Direction.NONE;
-            }
-
-            public override void move()
-            {
-                switch (m_direction)
-                {
-                    case Direction.DOWN:
-                        y += 40;
-                        break;
-                    case Direction.UP:
-                        y -= 40;
-                        break;
-                    case Direction.LEFT:
-                        x -= 40;
-                        break;
-                    case Direction.RIGHT:
-                        x += 40;
-                        break;
-                }
-            }
-        }
-
-        public class BodyPart : PositionedEntity
-        {
-            PositionedEntity m_next;
-            public BodyPart(PositionedEntity next)
-                : base(next.x, next.y, 40, 40, "pack://application:,,,/Resources/body.png")
-            {
-                m_next = next;
-            }
-
-            public override void move()
-            {
-                x = m_next.x;
-                y = m_next.y;
-            }
-        }
     }
 }
